+++ .agentplane/tasks/202608261249-BXQZ97/pr/diffstat.txt
@@ -1,7 +1,8 @@
  .../src/commands/pr/provider-update-branch.test.ts | 230 +++++++++++++
  .../src/commands/pr/provider-update-branch.ts      | 368 +++++++++++++++++++++
+ .../shared/provider-update-branch-route.ts         |  68 ++++
  .../route-decision-blockers.quality-review.test.ts |  91 +++++
- .../src/commands/shared/route-decision-blockers.ts |  66 ++++
+ .../src/commands/shared/route-decision-blockers.ts |   5 +
  .../src/commands/shared/route-gate-priority.ts     |   3 +-
  .../agentplane/src/commands/shared/route-oracle.ts |   1 +
  .../commands/shared/side-effect-authority.test.ts  |   4 +
@@ -11,10 +12,11 @@
  .../workflow-operation-projection.registry.test.ts |  49 +++
  .../shared/workflow-operation-projection.ts        |  33 ++
  .../src/commands/shared/workflow-postconditions.ts |   6 +
- .../src/commands/shared/workflow-step-branch.ts    |  22 ++
+ .../src/commands/shared/workflow-step-branch.ts    |   9 +-
  ...rkflow-step-projections.conflict-rework.test.ts | 112 +++++++
- .../src/commands/shared/workflow-step.ts           |  31 ++
+ .../shared/workflow-step-provider-update-branch.ts |  41 +++
+ .../src/commands/shared/workflow-step.ts           |   6 +
  .../task/branch-task-supervisor-operations.test.ts | 107 ++++++
  .../task/branch-task-supervisor-operations.ts      |  33 ++
  .../src/commands/task/configured-authority.test.ts |   1 +
- 19 files changed, 1159 insertions(+), 1 deletion(-)
+ 21 files changed, 1165 insertions(+), 5 deletions(-)
