+++ .agentplane/tasks/202608261249-BXQZ97/pr/review.md
@@ -12,8 +12,8 @@ Created: 2026-08-26T13:12:22.036Z
 
 ## Verification
 
-- State: pending
-- Note: Not recorded yet.
+- State: needs_rework
+- Note: Rework: Declared check failed: bun run ci:local:full
 - Canonical workflow state lives in the task README.
 
 ## Handoff Notes
@@ -31,8 +31,9 @@ Created: 2026-08-26T13:12:22.036Z
 ```text
  .../src/commands/pr/provider-update-branch.test.ts | 230 +++++++++++++
  .../src/commands/pr/provider-update-branch.ts      | 368 +++++++++++++++++++++
+ .../shared/provider-update-branch-route.ts         |  68 ++++
  .../route-decision-blockers.quality-review.test.ts |  91 +++++
- .../src/commands/shared/route-decision-blockers.ts |  66 ++++
+ .../src/commands/shared/route-decision-blockers.ts |   5 +
  .../src/commands/shared/route-gate-priority.ts     |   3 +-
  .../agentplane/src/commands/shared/route-oracle.ts |   1 +
  .../commands/shared/side-effect-authority.test.ts  |   4 +
@@ -42,13 +43,14 @@ Created: 2026-08-26T13:12:22.036Z
  .../workflow-operation-projection.registry.test.ts |  49 +++
  .../shared/workflow-operation-projection.ts        |  33 ++
  .../src/commands/shared/workflow-postconditions.ts |   6 +
- .../src/commands/shared/workflow-step-branch.ts    |  22 ++
+ .../src/commands/shared/workflow-step-branch.ts    |   9 +-
  ...rkflow-step-projections.conflict-rework.test.ts | 112 +++++++
- .../src/commands/shared/workflow-step.ts           |  31 ++
+ .../shared/workflow-step-provider-update-branch.ts |  41 +++
+ .../src/commands/shared/workflow-step.ts           |   6 +
  .../task/branch-task-supervisor-operations.test.ts | 107 ++++++
  .../task/branch-task-supervisor-operations.ts      |  33 ++
  .../src/commands/task/configured-authority.test.ts |   1 +
- 19 files changed, 1159 insertions(+), 1 deletion(-)
+ 21 files changed, 1165 insertions(+), 5 deletions(-)
 ```
 
 </details>
