+++ .agentplane/tasks/202608261249-BXQZ97/verification/20260826140354671-f118e6b643ac2277.json
@@ -1,0 +1,132 @@
+{
+  "details": "Command: bunx vitest run packages/agentplane/src/commands/pr\nResult: pass\nEvidence: .agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-1\nScope: branch_pr task 202608261249-BXQZ97 declared verification\n\nCommand: bunx vitest run packages/agentplane/src/commands/shared/route-decision-blockers.quality-review.test.ts packages/agentplane/src/commands/shared/workflow-step-projections.test.ts packages/agentplane/src/commands/shared/workflow-operation-projection.registry.test.ts\nResult: pass\nEvidence: .agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-2\nScope: branch_pr task 202608261249-BXQZ97 declared verification\n\nCommand: bunx vitest run packages/agentplane/src/commands/task/branch-task-supervisor-operations.test.ts\nResult: pass\nEvidence: .agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-3\nScope: branch_pr task 202608261249-BXQZ97 declared verification\n\nCommand: bun run ci:local:full\nResult: fail\nEvidence: .agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-4\nScope: branch_pr task 202608261249-BXQZ97 declared verification",
+  "digest": "sha256:f118e6b643ac227789ec13091f28011fbbda293fdc105ed52046007651f35db9",
+  "implementation_sha": "c2138ac88b0d44ad1bfb6ed2dea6f1e2d47efe04",
+  "input": {
+    "context": {
+      "digest": "sha256:45eb23ef20a43f6fc625b415551b2ccfff5623952b40648a24cc9432769f384f",
+      "paths": [
+        ".agentplane/WORKFLOW.md",
+        ".github/ISSUE_TEMPLATE/config.yml",
+        "bun.lock",
+        "depcruise.config.cjs",
+        "eslint.config.cjs",
+        "knip.json",
+        "package.json",
+        "packages/agentplane/package.json",
+        "packages/agentplane/src/cli/run-cli/commands/config.ts",
+        "packages/agentplane/src/runner/config.test.ts",
+        "packages/agentplane/src/runner/config.ts",
+        "packages/agentplane/tsconfig.json",
+        "packages/agentplane/tsup.config.ts",
+        "packages/core/package.json",
+        "packages/core/schemas/config.schema.json",
+        "packages/core/src/config/config.test.ts",
+        "packages/core/src/config/config.ts",
+        "packages/core/tsconfig.json",
+        "packages/core/tsup.config.ts",
+        "packages/recipes/package.json",
+        "packages/recipes/tsconfig.json",
+        "packages/recipes/tsup.config.ts",
+        "packages/spec/examples/config.json",
+        "packages/spec/package.json",
+        "packages/spec/schemas/config.schema.json",
+        "packages/testkit/package.json",
+        "packages/testkit/tsconfig.json",
+        "schemas/config.schema.json",
+        "tsconfig.json",
+        "turbo.json",
+        "vitest.config.ts",
+        "vitest.workspace.ts",
+        "website/bun.lock",
+        "website/docusaurus.config.ts",
+        "website/package.json",
+        "website/tsconfig.json"
+      ]
+    },
+    "digest": "sha256:8cdd6d46b7126ef0ccdc977f412b956e5a89d7a471a94b91a72fcb1a17621e2d",
+    "environment": {
+      "digest": "sha256:5eaa920b58022e097fc5b5afbe1e8caca7cdf6d6ad34230c2ac8eadaf77c86f9",
+      "runtime": {
+        "architecture": "arm64",
+        "bun_major": null,
+        "node_major": "26",
+        "platform": "darwin"
+      }
+    },
+    "evidence": {
+      "details_digest": "sha256:c3bc6647a1e3459a16a66b2b2fc87c8c618f98b0974f4aabf0a93360ad76afed",
+      "digest": "sha256:96362bb8df8577d68cc3173145d5022e292b13675925be8a446ec6c37a891cb7",
+      "references": [
+        {
+          "digest": "sha256:91648fcf04cafec378386ac5e80caa58ead8d570a0ef8ffdba7e26bf5d1d5af4",
+          "fragment": "check-1",
+          "path": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-1",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:91648fcf04cafec378386ac5e80caa58ead8d570a0ef8ffdba7e26bf5d1d5af4",
+          "fragment": "check-2",
+          "path": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-2",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:91648fcf04cafec378386ac5e80caa58ead8d570a0ef8ffdba7e26bf5d1d5af4",
+          "fragment": "check-3",
+          "path": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-3",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:91648fcf04cafec378386ac5e80caa58ead8d570a0ef8ffdba7e26bf5d1d5af4",
+          "fragment": "check-4",
+          "path": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608261249-BXQZ97/supervision/declared-checks.json#check-4",
+          "source": "filesystem"
+        }
+      ]
+    },
+    "execution": {
+      "authoritative_task_source": "task_worktree",
+      "base_ref": "79bc13ff33358c49e216901f59c8fbc0a17987d2",
+      "base_sha": "79bc13ff33358c49e216901f59c8fbc0a17987d2",
+      "digest": "sha256:bf64a25deaa3f18cb7d367e5188b74e3ca14992c259247b8988ddb7d99ff3319",
+      "primary_task_id": "202608261249-BXQZ97",
+      "reason_codes": [
+        "agent_preferred_branch_pr",
+        "effect_external_write",
+        "repository_branch_pr_floor",
+        "reversibility_recovery_required"
+      ],
+      "repository_mode": "branch_pr",
+      "requested_mode": "branch_pr",
+      "route_source": "execution_contract",
+      "selected_mode": "branch_pr",
+      "task_ids": [
+        "202608261249-BXQZ97"
+      ]
+    },
+    "implementation": {
+      "base_sha": "79bc13ff33358c49e216901f59c8fbc0a17987d2",
+      "digest": "sha256:c8dae5d866c61a46714d34be87fde1e130ee2a5574851b5a8e72f001dda317bb",
+      "strategy": "branch_diff",
+      "target_sha": "c2138ac88b0d44ad1bfb6ed2dea6f1e2d47efe04"
+    },
+    "kind": "task_verification_input",
+    "schema_version": 4,
+    "verification_contract_digest": "sha256:6c55246ef558944022c7abfe43167a396b03b1717c06e8cc71a5faf196ad7478",
+    "verify_steps_digest": "sha256:24bffea4cd69fff4560e7111e963b698310685c305463625baa43d9cac58eb99"
+  },
+  "kind": "task_verification_record",
+  "note": "Rework: Declared check failed: bun run ci:local:full",
+  "recorded_at": "2026-08-26T14:03:54.671Z",
+  "result": "needs_rework",
+  "schema_version": 2,
+  "scope": "PLANNER fallback scaffold for \"Add a digest-bound provider update-branch recovery transition for stale hosted PR heads\". Replace with task-specific acceptance checks when PLANNER context is available.\n\n1. Review the requested outcome for \"Add a digest-bound provider update-branch recovery transition for stale hosted PR heads\". Expected: the visible result matches ## Summary and stays inside approved scope.\n2. Run the most relevant validation step for this task. Expected: it succeeds without unexpected regressions in touched behavior.\n3. Compare the final result against ## Scope and record any residual follow-up in ## Findings. Expected: open edges are explicit rather than implicit.",
+  "scope_digest": "sha256:24bffea4cd69fff4560e7111e963b698310685c305463625baa43d9cac58eb99",
+  "task_id": "202608261249-BXQZ97",
+  "verification_command": "agentplane verify 202608261249-BXQZ97 --rework --by SUPERVISOR",
+  "verifier": "SUPERVISOR"
+}
