+++ packages/agentplane/src/commands/shared/workflow-step.ts
@@ -4,14 +4,15 @@ import type { TaskRepositoryEffect } from "@agentplaneorg/core/tasks";
 import type { TaskData } from "../../backends/task-backend.js";
 import type { PrFlowStatusReport } from "../pr/flow-status.js";
 import type { ConflictReworkPreparation } from "../pr/conflict-rework.js";
-import type { GitHostIdentity } from "../pr/internal/git-host-identity.js";
 import type { TaskResumeContext } from "../task/handoff.shared.js";
+import type { ProviderUpdateBranchOperationParams } from "./provider-update-branch-route.js";
 import type { RouteBatchOwnership } from "./route-batch-ownership.js";
 import type { RouteCleanupProbe, RouteNextAction } from "./route-decision-types.js";
 import type { RouteBlocker, RouteExecutionPacket, RouteOracle } from "./route-oracle.js";
 import type { TaskWorktreeCleanliness } from "./task-worktree-cleanliness.js";
 import type { ForeignTaskReadmeReplicaRepair } from "./task-worktree-foreign-artifact-repair.js";
 import { foreignTaskReadmeReplicaRepairOperation } from "./workflow-step-foreign-task-readme-repair.js";
+import { PROVIDER_UPDATE_BRANCH_OPERATION_SPEC } from "./workflow-step-provider-update-branch.js";
 import { POSTCONDITION, type WorkflowPostcondition } from "./workflow-postconditions.js";
 
 export type WorkflowRole = RouteExecutionPacket["recommendedRole"];
@@ -106,15 +107,7 @@ export type WorkflowOperationParams = {
   "pr.open": { taskId: string; author: string; includeTaskIds: readonly string[] };
   "pr.sync_or_verify": { taskId: string; includeTaskIds: readonly string[] };
   "provider.pr.refresh": { taskId: string };
-  "provider.pr.update_branch": {
-    taskId: string;
-    identity: GitHostIdentity;
-    prNumber: number;
-    branch: string;
-    baseBranch: string;
-    expectedHeadSha: string;
-    expectedBaseSha: string;
-  };
+  "provider.pr.update_branch": ProviderUpdateBranchOperationParams;
   "flow.repair.foreign_task_readme": { taskId: string };
   "route.remote.refresh": { taskId: string };
   "runner.follow":
@@ -335,25 +328,7 @@ export const WORKFLOW_OPERATION_REGISTRY = {
     verificationCandidate: "agentplane pr flow status <task-id>",
     needsVerificationRecord: false,
   },
-  "provider.pr.update_branch": {
-    type: "provider_update_branch",
-    phase: "provider_pr_update_branch_required",
-    checkout: "task_worktree",
-    role: "INTEGRATOR",
-    expectedPostconditions: [
-      POSTCONDITION.providerBranchUpdated,
-      POSTCONDITION.providerObserved,
-      POSTCONDITION.routeRecomputed,
-    ],
-    mustNot: [
-      "do not update a different PR, branch, base, provider identity, or head than the exact approved operation",
-      "do not repeat an effect-in-doubt update; reconcile fresh provider ancestry before another mutation",
-      "do not merge, rebase, force-push, or edit AgentPlane task state as a substitute for provider update-branch",
-    ],
-    triggersGitHooks: false,
-    verificationCandidate: "agentplane task next-action <task-id> --remote --explain",
-    needsVerificationRecord: false,
-  },
+  "provider.pr.update_branch": PROVIDER_UPDATE_BRANCH_OPERATION_SPEC,
   "route.remote.refresh": {
     type: "provider_refresh",
     phase: "remote_route_refresh_needed",
